feat:MNST-12 finished whole logic in all package and changes in pom.xml

+++ src/main/java/com/spring/boot/project/ms/manga/store/domain/model/cart/CartItem.java
@@ -1,7 +1,9 @@
 package com.spring.boot.project.ms.manga.store.domain.model.cart;
 
+import java.math.BigDecimal;
+
 public record CartItem(
         String isbn,
         int quantity,
-        double price) {
+        BigDecimal price) {
 }
